Update menu types, remove unused utils
